Added load screen to app
