+++ app/src/main/java/com/example/mayank/tradex/Title_TradeX.java
@@ -1,6 +1,8 @@
 package com.example.mayank.tradex;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
 import android.support.v7.app.AppCompatActivity;
@@ -8,37 +10,24 @@ import android.view.MenuItem;
 import android.widget.TextView;
 
 public class Title_TradeX extends AppCompatActivity {
+    private static int SPLASH_TIME_OUT = 4000;
 
-    private TextView mTextMessage;
-
-    private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener
-            = new BottomNavigationView.OnNavigationItemSelectedListener() {
-
-        @Override
-        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-            switch (item.getItemId()) {
-                case R.id.navigation_home:
-                    mTextMessage.setText(R.string.title_home);
-                    return true;
-                case R.id.navigation_dashboard:
-                    mTextMessage.setText(R.string.title_dashboard);
-                    return true;
-                case R.id.navigation_notifications:
-                    mTextMessage.setText(R.string.title_notifications);
-                    return true;
-            }
-            return false;
-        }
-    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_title__trade_x);
 
-        mTextMessage = (TextView) findViewById(R.id.message);
-        BottomNavigationView navigation = (BottomNavigationView) findViewById(R.id.navigation);
-        navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
+        new Handler().postDelayed (new Runnable() {
+            @Override
+            public void run() {
+                Intent loadintent = new Intent(Title_TradeX.this, LoadScreenActivity.class);
+                startActivity(loadintent);
+                finish();
+            }
+        }, SPLASH_TIME_OUT);
+
+
     }
 
 }
